+++ CHANGELOG.md
@@ -8,6 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.19.0 - 2023-03-13
+
+### Added
+* Right sidebar
+
+### Changed
+* Create trigger via Right sidebar
+
+
 ## 0.18.0 - 2023-03-10
 
 ### Added